Remove package reference, remove proto files, replace logs and metrics
