+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/ExportClient/IExportClient.cs
@@ -4,17 +4,17 @@
 namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation.ExportClient;
 
 /// <summary>Export client interface.</summary>
-/// <typeparam name="TRequest">Type of export request.</typeparam>
-internal interface IExportClient<in TRequest>
+internal interface IExportClient
 {
     /// <summary>
     /// Method for sending export request to the server.
     /// </summary>
-    /// <param name="request">The request to send to the server.</param>
+    /// <param name="buffer">The request body to send to the server.</param>
+    /// <param name="contentLength">length of the content.</param>
     /// <param name="deadlineUtc">The deadline time in utc for export request to finish.</param>
     /// <param name="cancellationToken">An optional token for canceling the call.</param>
     /// <returns><see cref="ExportClientResponse"/>.</returns>
-    ExportClientResponse SendExportRequest(TRequest request, DateTime deadlineUtc, CancellationToken cancellationToken = default);
+    ExportClientResponse SendExportRequest(byte[] buffer, int contentLength, DateTime deadlineUtc, CancellationToken cancellationToken = default);
 
     /// <summary>
     /// Method for shutting down the export client.
