--- src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/Transmission/ProtobufOtlpExporterRetryTransmissionHandler.cs
@@ -1,35 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-using OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation.ExportClient;
-
-namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation.Transmission;
-
-internal sealed class ProtobufOtlpExporterRetryTransmissionHandler : ProtobufOtlpExporterTransmissionHandler
-{
-    internal ProtobufOtlpExporterRetryTransmissionHandler(IProtobufExportClient exportClient, double timeoutMilliseconds)
-        : base(exportClient, timeoutMilliseconds)
-    {
-    }
-
-    protected override bool OnSubmitRequestFailure(byte[] request, int contentLength, ExportClientResponse response)
-    {
-        var nextRetryDelayMilliseconds = OtlpRetry.InitialBackoffMilliseconds;
-        while (RetryHelper.ShouldRetryRequest(response, nextRetryDelayMilliseconds, out var retryResult))
-        {
-            // Note: This delay cannot exceed the configured timeout period for otlp exporter.
-            // If the backend responds with `RetryAfter` duration that would result in exceeding the configured timeout period
-            // we would fail fast and drop the data.
-            Thread.Sleep(retryResult.RetryDelay);
-
-            if (this.TryRetryRequest(request, contentLength, response.DeadlineUtc, out response))
-            {
-                return true;
-            }
-
-            nextRetryDelayMilliseconds = retryResult.NextRetryDelayMilliseconds;
-        }
-
-        return false;
-    }
-}
