--- src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/opentelemetry/proto/collector/trace/v1/trace_service_http.yaml
@@ -1,9 +1,0 @@
-# This is an API configuration to generate an HTTP/JSON -> gRPC gateway for the
-# OpenTelemetry service using github.com/grpc-ecosystem/grpc-gateway.
-type: google.api.Service
-config_version: 3
-http:
- rules:
- - selector: opentelemetry.proto.collector.trace.v1.TraceService.Export
-   post: /v1/traces
-   body: "*"
